requirements.txt
tabulate
ulid-py
argcomplete
alembic==1.3.1
Mako==1.1.0
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-editor==1.0.4
six==1.13.0
SQLAlchemy==1.3.11
